+++ harfbuzz/src/hb-ot-face-table-list.hh
@@ -59,9 +59,6 @@ HB_OT_ACCELERATOR (OT, post)
 #ifndef HB_NO_STAT
 HB_OT_TABLE (OT, STAT)
 #endif
-#ifndef HB_NO_META
-HB_OT_ACCELERATOR (OT, meta)
-#endif
 
 /* Vertical layout. */
 HB_OT_TABLE (OT, vhea)